move get-ip-address
