+++ packages/utils/package.json
@@ -4,8 +4,7 @@
   "private": true,
   "type": "module",
   "exports": {
-    ".": "./src/index.ts",
-    "./get-ip-address": "./src/get-ip-address.ts"
+    ".": "./src/index.ts"
   },
   "scripts": {
     "lint": "eslint .",
@@ -13,12 +12,8 @@
     "test": "vitest run",
     "type-check": "tsc --noEmit"
   },
-  "optionalDependencies": {
-    "is-ip": "^5.0.1"
-  },
   "devDependencies": {
     "@oyster/eslint-config": "*",
-    "is-ip": "^5.0.1",
     "nanoid": "^3.0.0",
     "vitest": "^1.3.1"
   }
